Transform the project into a generic CLI tool
This lets you invoke the tool manually outside a GitHub Actions context.
I still intend to wrap it in a reusable GitHub Actions later, as this is
the original goal of the project.

Dropping the `npm` part of the project name signifies that you can use
the tool for non-npm releases as well, such as other reusable GitHub
Actions projects.